+++ service/history/service.go
@@ -28,6 +28,7 @@ import (
 	"github.com/uber/cadence/common/log/tag"
 	"github.com/uber/cadence/common/metrics"
 	"github.com/uber/cadence/common/persistence"
+	espersistence "github.com/uber/cadence/common/persistence/elasticsearch"
 	persistencefactory "github.com/uber/cadence/common/persistence/persistence-factory"
 	"github.com/uber/cadence/common/service"
 	"github.com/uber/cadence/common/service/config"
@@ -303,7 +304,7 @@ func (s *Service) Start() {
 		if err != nil {
 			log.Fatal("Creating visibility producer failed", tag.Error(err))
 		}
-		esVisibility = persistence.NewESVisibilityManager("", nil, nil, visibilityProducer,
+		esVisibility = espersistence.NewESVisibilityManager("", nil, nil, visibilityProducer,
 			s.metricsClient, log)
 	}
 	visibility = persistence.NewVisibilityManagerWrapper(visibility, esVisibility, dynamicconfig.GetBoolPropertyFnFilteredByDomain(false))
